+++ instrumentation/logback/logback-mdc-1.0/javaagent/src/addBaggageTest/java/io/opentelemetry/javaagent/instrumentation/logback/v1_0/LogbackWithBaggageTest.java
@@ -5,9 +5,9 @@
 
 package io.opentelemetry.javaagent.instrumentation.logback.v1_0;
 
-import io.opentelemetry.instrumentation.logback.mdc.v1_0.AbstractLogbackAgentTest;
+import io.opentelemetry.instrumentation.logback.mdc.v1_0.AbstractLogbackTest;
 
-class LogbackWithBaggageTest extends AbstractLogbackAgentTest {
+class LogbackWithBaggageTest extends AbstractLogbackTest {
   @Override
   protected boolean expectBaggage() {
     return true;
